Cleaning up.

+++ src/std/str.c
@@ -5,7 +5,7 @@ int strlen(char* f){
   return ct;
 }
 
-char* itoa(uint32_t num, int base){ 
+char* itoa(uint64_t num, int base){ 
 	static char repr[]= "0123456789abcdef";
 	static char buffer[50]; 
 	char *ptr;
